requirements.txt
# ================================
# Core API
# ================================
fastapi>=0.110
uvicorn>=0.27
starlette>=0.36

# ================================
# Data validation
# ================================
pydantic>=2.6

# ================================
# PostgreSQL + pgvector
# ================================
psycopg2-binary>=2.9
sqlalchemy>=2.0
langchain-postgres>=0.0.6
pgvector>=0.2.5

# ================================
# LangChain core
# ================================
langchain-core>=0.1.45
langchain-huggingface>=0.0.8

# ================================
# Embeddings / ML
# ================================
sentence-transformers>=2.5
torch>=2.0
transformers>=4.39
accelerate>=0.27

# ================================
# HuggingFace utilities (REQUIRED)
# ================================
huggingface-hub>=0.21

# ================================
# Lite model (GGUF / CPU)
# ================================
llama-cpp-python>=0.2.20

# ================================
# Redis (memory / debug)
# ================================
redis>=5.0

# ================================
# File uploads
# ================================
python-multipart>=0.0.9

# ================================
# Document parsing (PDF ingestion)
# ================================
unstructured[pdf]>=0.12.6

# ================================
# Token counting (RAG metadata)
# ================================
tiktoken>=0.6.0

# ================================
# Utilities
# ================================
numpy>=1.24
tqdm>=4.66
requests>=2.31

# ================================
# Stability (recommended)
# ================================
protobuf>=4.25
